K9 toss and fetch class
I've very coarsly created a function for opening K9 tossnfectch files that I've exported from tableau. Currently it only includes functionf for opening the file, and getting the data by the season. My next mission is to add functions for cleaning the data, and choosing data on columns other that the season. Most of these functions will probably just be wrapping already existing pandas functions, but I think in the end it will make processing the data more streamlined